--- src/hooks/useProfitAndLoss/index.tsx
@@ -1,3 +1,0 @@
-export { useProfitAndLoss } from './useProfitAndLoss'
-export { useProfitAndLossQuery } from './useProfitAndLossQuery'
-export { useProfitAndLossLTM } from './useProfitAndLossLTM'
